+++ tsdb/engine/tsm1/engine.go
@@ -330,6 +330,8 @@ func (e *Engine) SetLogOutput(w io.Writer) {
 
 // LoadMetadataIndex loads the shard metadata into memory.
 func (e *Engine) LoadMetadataIndex(shardID uint64, index *tsdb.DatabaseIndex) error {
+	now := time.Now()
+
 	// Save reference to index for iterator creation.
 	e.index = index
 
@@ -364,6 +366,7 @@ func (e *Engine) LoadMetadataIndex(shardID uint64, index *tsdb.DatabaseIndex) er
 		}
 	}
 
+	e.traceLogger.Printf("Meta data index for shard %d loaded in %v", shardID, time.Since(now))
 	return nil
 }
 
@@ -691,6 +694,7 @@ func (e *Engine) WriteSnapshot() error {
 	defer func() {
 		if started != nil {
 			e.Cache.UpdateCompactTime(time.Now().Sub(*started))
+			e.logger.Printf("Snapshot for path %s written in %v", e.path, time.Since(*started))
 		}
 	}()
 
@@ -725,7 +729,9 @@ func (e *Engine) WriteSnapshot() error {
 	// The snapshotted cache may have duplicate points and unsorted data.  We need to deduplicate
 	// it before writing the snapshot.  This can be very expensive so it's done while we are not
 	// holding the engine write lock.
+	dedup := time.Now()
 	snapshot.Deduplicate()
+	e.traceLogger.Printf("Snapshot for path %s deduplicated in %v", e.path, time.Since(dedup))
 
 	return e.writeSnapshotAndCommit(closedFiles, snapshot)
 }
@@ -789,6 +795,7 @@ func (e *Engine) compactCache() {
 			e.Cache.UpdateAge()
 			if e.ShouldCompactCache(e.WAL.LastWriteTime()) {
 				start := time.Now()
+				e.traceLogger.Printf("Compacting cache for %s", e.path)
 				err := e.WriteSnapshot()
 				if err != nil && err != errCompactionsDisabled {
 					e.logger.Printf("error writing snapshot: %v", err)
@@ -987,6 +994,7 @@ func (e *Engine) compactTSMFull() {
 
 // reloadCache reads the WAL segment files and loads them into the cache.
 func (e *Engine) reloadCache() error {
+	now := time.Now()
 	files, err := segmentFileNames(e.WAL.Path())
 	if err != nil {
 		return err
@@ -1006,6 +1014,7 @@ func (e *Engine) reloadCache() error {
 		return err
 	}
 
+	e.traceLogger.Printf("Reloaded WAL cache %s in %v", e.WAL.Path(), time.Since(now))
 	return nil
 }